Migrate contents of petal.py into petal/__init__.py for organization
